refactor: Change deposit points and weight data type to float
Updates the data type for deposit points and weight from integer to float across the navigation graph and related screens.

- In `AuthenticatedNavHost`, the `navArgument` types for the `SuccessDeposit` route are changed from `NavType.IntType` to `NavType.FloatType`.
- In `SuccessDepositScreen`, the `points` and `weight` parameters are updated from `Int` to `Float` to correctly handle the floating-point values passed through navigation.

+++ app/src/main/java/com/ahargunyllib/growth/model/Collection.kt
@@ -13,7 +13,7 @@ data class Collection (
     val id: String = "",
     val userId: String = "",
     val partnerId: String = "",
-    val totalWeightKg: Int = 0,
+    val totalWeightKg: Float = 0F,
     val receivedPoints: Int = 0,
     val status: CollectionStatus = CollectionStatus.PENDING,
     val createdAt: String = ""
